didn't compiled with flags

+++ ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: oozsertt <oozsertt@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/09/14 17:56:51 by user42            #+#    #+#             */
-/*   Updated: 2020/09/17 15:19:46 by oozsertt         ###   ########.fr       */
+/*   Updated: 2020/09/18 01:03:08 by oozsertt         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	while (s1[i] == s2[i] && i < n)
